+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note: Go into settings to make sure it is correct.
 * Added keybindings for inviting and scanning (Found in settings, Invite Settings).
 Note: This does not overwrite any existing keybindings in WoW.
 * Players in an instance will not be messaged or invited.
+* Fixed issue with welcome message not using GM settings.
 
 
 ## 2.1.36